requirements.txt
yt-dlp
faster-whisper
imageio-ffmpeg
PyYAML
markitdown[all]
opendataloader-pdf
openai
python-dotenv

# Optional: GPU acceleration (NVIDIA CUDA)
# pip install faster-whisper[cuda]
